fixed the data type of the tag varaible.

+++ result_process.h
@@ -5,7 +5,7 @@
 #define _RESULT_PROCESS
 
 typedef struct _Op_result {
-  u16 tag;
+  u8 tag;
   u64 time_spent;
 } Op_result;
